Add 1stDibs archive import script + theme taxonomy helper
scripts/import-archive.ts reconciles the artworks DB with the
1stDibs picks CSV. For each row:
  - SKU exists in DB:  upsert theme tags only (no metadata clobber)
                       + flag the row in the log if it looks partial
                       (no AI alt or non-R2 image — likely from a
                       prior interrupted run)
  - SKU not in DB:     insert artwork from CSV metadata, download
                       image, upload R2 4-variant pipeline, generate
                       AI alt text via Claude Vision, attach themes.
                       on_website=true per spec.

Themes are a controlled vocabulary of 8 (music, people, plants,
animals, abstract, other, food, pop culture) stored in the existing
categories table with the new kind='theme' discriminator. Theme
normalization (clir/clear prefix stripping, case folding, dedup,
fixed-set validation) is in src/lib/themes.ts so Project B (theme +
decade UI dropdowns) can reuse it. node:test covers the helper.

Concurrency uses cache + single-flight Promise gating + Supabase
upsert(onConflict) for both theme categories and artists, so 3
parallel workers don't race on the unique slug constraint. Log
writing wrapped in try/finally so a worker rejection still flushes
the log to disk.

Out of scope per spec: image validation for existing artworks,
inactive marking for DB-only SKUs, metadata refresh on existing
rows.

Schema migration to add categories.kind was applied manually before
this commit. supabase/migrations/001_initial.sql synced to match.

Plan also got a small fix: the original "verify imports resolve"
smoke test used a dynamic import that triggered main() and accidentally
started writing to the DB. Replaced with a safer "run without env
vars and confirm it exits at the env-var check" approach.
