+++ package.json
@@ -13,6 +13,7 @@
     "seed:categories": "tsx scripts/seed-categories.ts",
     "generate:descriptions": "tsx scripts/generate-descriptions.ts",
     "import:descriptions": "tsx --env-file=.env.local scripts/import-human-descriptions.ts",
+    "import:archive": "tsx --env-file=.env.local scripts/import-archive.ts",
     "migrate:all": "tsx --env-file=.env.local scripts/migrate-and-describe.ts",
     "db:migrate": "tsx scripts/run-migration.ts"
   },
