+++ scripts/import-archive.ts
@@ -1,210 +1,3 @@
-# Archive Import & Theme Taxonomy Implementation Plan
-
-> **For agentic workers:** REQUIRED SUB-SKILL: Use superpowers:subagent-driven-development (recommended) or superpowers:executing-plans to implement this plan task-by-task. Steps use checkbox (`- [ ]`) syntax for tracking.
-
-**Goal:** Reconcile the artworks DB with `tmp/ADD TAGS TO CREATIVE GROWTH PUBLIC ARCHIVE - 1stdibs_clir_picks_2026-03-17.csv` — upsert theme tags for the 909 SKUs already in the DB and insert + AI-describe the 1,150 SKUs not yet present.
-
-**Architecture:** One main script (`scripts/import-archive.ts`) that pages through the CSV, branches per row on whether the SKU exists in the DB, and dispatches to a small set of helpers (theme normalization extracted as a unit-testable pure function; image download/resize/upload and Claude Vision call copied from `migrate-and-describe.ts` — the spec explicitly accepts duplication for this PR's scope). Concurrency, checkpointing, and idempotency follow the existing `migrate-and-describe.ts` pattern.
-
-**Tech Stack:** TypeScript, Node 20+, `tsx`, `csv-parse/sync`, `@supabase/supabase-js`, `@aws-sdk/client-s3`, `sharp`, `@anthropic-ai/sdk`, Node's built-in test runner.
-
-**Spec reference:** `docs/superpowers/specs/2026-04-23-archive-import-design.md`
-
----
-
-## File Structure
-
-**New files:**
-- `src/lib/themes.ts` — pure function `normalizeThemes(raw: string)` that takes col T's raw string and returns an array of valid theme slugs from the fixed set. Lives in `src/lib/` (not `scripts/lib/`) because Project B's UI work will need it too.
-- `scripts/test-themes.ts` — `node:test` assertions for `normalizeThemes`.
-- `scripts/import-archive.ts` — main script. Parses CSV, branches per-row, handles new-row insert pipeline, writes log artifact.
-
-**Modified files:**
-- `package.json` — add `import:archive` npm script.
-- `supabase/migrations/001_initial.sql` — sync source-controlled schema with the applied `kind` column on `categories` (already added by user via SQL Editor).
-
----
-
-## Phase 0: Pre-flight
-
-### Task 0: Verify the `kind` column exists on `categories`
-
-**Files:** none (read-only verification)
-
-- [ ] **Step 1: Confirm column exists**
-
-Run:
-```bash
-npx tsx --env-file=.env.local -e "
-import { createClient } from '@supabase/supabase-js';
-const c = createClient(process.env.NEXT_PUBLIC_SUPABASE_URL, process.env.SUPABASE_SERVICE_ROLE_KEY);
-c.from('categories').select('id, name, kind').limit(3).then(({data, error}) => {
-  if (error) { console.error('SCHEMA MISMATCH:', error.message); process.exit(1); }
-  console.log('OK - kind column exists. Sample:');
-  data.forEach(r => console.log('  ', JSON.stringify(r)));
-});
-"
-```
-
-Expected: prints `OK - kind column exists.` followed by 3 sample category rows. The existing categories should show `kind: 'format'` (per the spec's backfill SQL).
-
-If this fails: stop. The user needs to re-run the SQL from the spec's "Schema changes" section.
-
----
-
-## Phase 1: Code
-
-### Task 1: TDD theme normalization — failing test first
-
-**Files:**
-- Create: `scripts/test-themes.ts`
-
-- [ ] **Step 1: Write the test file**
-
-Create `scripts/test-themes.ts`:
-
-```typescript
-import { strict as assert } from "node:assert";
-import { test } from "node:test";
-import { normalizeThemes, VALID_THEMES } from "../src/lib/themes";
-
-test("strips 'clir' prefix and returns the bare slug", () => {
-  assert.deepEqual(normalizeThemes("clir abstract"), ["abstract"]);
-});
-
-test("strips 'clear' prefix as well as 'clir'", () => {
-  assert.deepEqual(normalizeThemes("clear plants"), ["plants"]);
-});
-
-test("accepts comma-separated multi-theme strings", () => {
-  assert.deepEqual(
-    normalizeThemes("clir abstract, clir people"),
-    ["abstract", "people"]
-  );
-});
-
-test("trims whitespace and lowercases", () => {
-  assert.deepEqual(normalizeThemes("  CLIR Abstract  "), ["abstract"]);
-});
-
-test("preserves multi-word themes ('pop culture')", () => {
-  assert.deepEqual(normalizeThemes("clir pop culture"), ["pop culture"]);
-});
-
-test("returns empty array for empty input", () => {
-  assert.deepEqual(normalizeThemes(""), []);
-  assert.deepEqual(normalizeThemes("   "), []);
-});
-
-test("dedupes within a single row", () => {
-  assert.deepEqual(
-    normalizeThemes("clir abstract, clear abstract"),
-    ["abstract"]
-  );
-});
-
-test("drops values not in the fixed VALID_THEMES set", () => {
-  // 'date tag' style values should not survive
-  assert.deepEqual(normalizeThemes("clir 1980s"), []);
-  assert.deepEqual(normalizeThemes("clir music, clir bogus"), ["music"]);
-});
-
-test("VALID_THEMES is exactly the 8 expected values", () => {
-  assert.deepEqual(
-    [...VALID_THEMES].sort(),
-    ["abstract", "animals", "food", "music", "other", "people", "plants", "pop culture"]
-  );
-});
-
-test("returns themes in the order they appear (after dedup)", () => {
-  assert.deepEqual(
-    normalizeThemes("clir music, clir abstract, clir people"),
-    ["music", "abstract", "people"]
-  );
-});
-```
-
-- [ ] **Step 2: Run the test and verify it fails**
-
-Run: `npx tsx --test scripts/test-themes.ts`
-Expected: All tests fail with `Error: Cannot find module '../src/lib/themes'`.
-
----
-
-### Task 2: Implement theme normalization
-
-**Files:**
-- Create: `src/lib/themes.ts`
-
-- [ ] **Step 1: Write the implementation**
-
-Create `src/lib/themes.ts`:
-
-```typescript
-/**
- * Theme taxonomy — a fixed set of curated subject tags imported from
- * the 1stDibs CSV. See:
- *   docs/superpowers/specs/2026-04-23-archive-import-design.md
- *
- * In the source CSV, themes appear as comma-separated strings with a
- * "clir " or "clear " prefix (the prefix is inconsistent in the data).
- * normalizeThemes strips the prefix, lowercases, dedupes, and drops
- * anything outside VALID_THEMES.
- */
-
-export const VALID_THEMES = new Set([
-  "music",
-  "people",
-  "plants",
-  "animals",
-  "abstract",
-  "other",
-  "food",
-  "pop culture",
-]);
-
-export function normalizeThemes(raw: string): string[] {
-  if (!raw) return [];
-  const seen = new Set<string>();
-  const result: string[] = [];
-
-  for (const piece of raw.split(",")) {
-    const trimmed = piece.trim().toLowerCase();
-    if (!trimmed) continue;
-
-    // Strip leading "clir " or "clear " prefix
-    const stripped = trimmed
-      .replace(/^clir\s+/, "")
-      .replace(/^clear\s+/, "");
-
-    if (!VALID_THEMES.has(stripped)) continue;
-    if (seen.has(stripped)) continue;
-
-    seen.add(stripped);
-    result.push(stripped);
-  }
-
-  return result;
-}
-```
-
-- [ ] **Step 2: Run the test and verify it passes**
-
-Run: `npx tsx --test scripts/test-themes.ts`
-Expected: All 10 tests pass.
-
----
-
-### Task 3: Implement the main import script
-
-**Files:**
-- Create: `scripts/import-archive.ts`
-
-- [ ] **Step 1: Write the script**
-
-Create `scripts/import-archive.ts`:
-
-```typescript
 #!/usr/bin/env npx tsx
 /**
  * import-archive.ts
@@ -316,11 +109,16 @@ interface CsvRow {
 interface DbArtwork {
   id: string;
   sku: string | null;
+  alt_text_long: string | null;
+  description_origin: string | null;
+  image_url: string | null;
 }
 
-interface ThemeCategory {
+interface DbArtworkInfo {
   id: string;
-  name: string;
+  alt_text_long: string | null;
+  description_origin: string | null;
+  image_url: string | null;
 }
 
 interface Progress {
@@ -520,6 +318,14 @@ async function resolveArtistId(artistName: string): Promise<string | null> {
 
 // ─── Main ─────────────────────────────────────────────────────────────────
 async function main() {
+  // Resumability caveat: if killed between Branch B's INSERT (line ~430) and
+  // the checkpoint write (~end of processRow), the next run will see the new
+  // SKU as already-existing, take Branch A, and skip the missing image/AI
+  // steps. Branch A flags any existing row that looks partial (no AI alt OR
+  // non-R2 image_url) in the log's `notes` column. After a crash, search the
+  // log for "WARNING:" and fix those rows manually (or delete them and re-run
+  // to let Branch B re-process from scratch).
+
   // 1. Parse CSV
   if (!fs.existsSync(CSV_PATH)) {
     console.error(`CSV not found: ${CSV_PATH}`);
@@ -536,17 +342,24 @@ async function main() {
 
   // 2. Load DB SKU set
   console.log("Fetching all artwork SKUs from DB...");
-  const dbSkuToId = new Map<string, string>();
+  const dbSkuToId = new Map<string, DbArtworkInfo>();
   let offset = 0;
   while (true) {
     const { data, error } = await supabase
       .from("artworks")
-      .select("id, sku")
+      .select("id, sku, alt_text_long, description_origin, image_url")
       .range(offset, offset + PAGE_SIZE - 1);
     if (error) { console.error(error); process.exit(1); }
     if (!data || data.length === 0) break;
     for (const a of data as DbArtwork[]) {
-      if (a.sku) dbSkuToId.set(a.sku.trim(), a.id);
+      if (a.sku) {
+        dbSkuToId.set(a.sku.trim(), {
+          id: a.id,
+          alt_text_long: a.alt_text_long,
+          description_origin: a.description_origin,
+          image_url: a.image_url,
+        });
+      }
     }
     if (data.length < PAGE_SIZE) break;
     offset += PAGE_SIZE;
@@ -560,7 +373,7 @@ async function main() {
 
   // 4. Build work queue
   const work = rows
-    .map((r, idx) => ({ idx, row: r, sku: (r.SKU || "").trim() }))
+    .map((r) => ({ row: r, sku: (r.SKU || "").trim() }))
     .filter(({ sku }) => sku && !doneSet.has(sku));
   console.log(`Remaining rows to process: ${work.length}`);
 
@@ -601,12 +414,28 @@ async function main() {
         .filter((p) => p && !VALID_THEMES.has(p));
       log.themes_dropped = [...new Set(droppedRaw)].join("; ");
 
-      const existingId = dbSkuToId.get(sku);
-      if (existingId) {
+      const existing = dbSkuToId.get(sku);
+      if (existing) {
         // Branch A: theme upsert only
         log.branch = "existing_themed";
-        log.artwork_id = existingId;
-        await attachThemes(existingId, themes);
+        log.artwork_id = existing.id;
+
+        // Surface partial-state rows from prior interrupted runs so the
+        // maintainer can spot-check. We don't repair the row here.
+        const warnings: string[] = [];
+        if (existing.description_origin === null && existing.alt_text_long === null) {
+          warnings.push(
+            "WARNING: existing row missing AI alt text — likely partial from a prior interrupted run; review manually"
+          );
+        }
+        if (existing.image_url && existing.image_url.includes("artcld.com")) {
+          warnings.push(
+            "WARNING: image_url not on R2 — may indicate prior interrupted run"
+          );
+        }
+        if (warnings.length > 0) log.notes = warnings.join("; ");
+
+        await attachThemes(existing.id, themes);
         log.themes_attached = themes.join("; ");
       } else {
         // Branch B: full insert + image + AI + themes
@@ -722,14 +551,18 @@ async function main() {
       await processRow(work[i]);
     }
   }
-  await Promise.all(Array.from({ length: CONCURRENCY }, () => worker()));
-
-  // 6. Write log
-  if (!fs.existsSync(TMP_DIR)) fs.mkdirSync(TMP_DIR, { recursive: true });
-  writeCsv(LOG_FILE,
-    ["sku", "branch", "themes_attached", "themes_dropped", "image_status", "ai_status", "artwork_id", "notes"],
-    logs
-  );
+  try {
+    await Promise.all(Array.from({ length: CONCURRENCY }, () => worker()));
+  } finally {
+    // Write the log even if a worker rejected
+    if (!fs.existsSync(TMP_DIR)) fs.mkdirSync(TMP_DIR, { recursive: true });
+    writeCsv(LOG_FILE,
+      ["sku", "branch", "themes_attached", "themes_dropped", "image_status", "ai_status", "artwork_id", "notes"],
+      logs
+    );
+    console.log(`\nLog: ${LOG_FILE}`);
+    console.log(`Checkpoint: ${PROGRESS_FILE}`);
+  }
 
   // 7. Summary
   const byBranch = logs.reduce<Record<string, number>>((acc, l) => {
@@ -738,302 +571,9 @@ async function main() {
   }, {});
   console.log("\n=== Summary ===");
   Object.entries(byBranch).forEach(([k, v]) => console.log(`  ${k.padEnd(18)} ${v}`));
-  console.log(`\nLog: ${LOG_FILE}`);
-  console.log(`Checkpoint: ${PROGRESS_FILE}`);
 }
 
 main().catch((err) => {
   console.error("Fatal:", err);
   process.exit(1);
 });
-```
-
-- [ ] **Step 2: Verify script's modules resolve (SAFE — no DB writes)**
-
-Run the script WITHOUT `--env-file=.env.local` so env vars are missing — main() will exit at the env-var check before any DB activity:
-
-```bash
-npx tsx scripts/import-archive.ts 2>&1 | head -5
-```
-
-Expected: prints `Error: Set NEXT_PUBLIC_SUPABASE_URL and SUPABASE_SERVICE_ROLE_KEY` and exits with code 1.
-
-This is sufficient to prove the file parses and all imports resolve (`csv-parse`, `@supabase/supabase-js`, `@aws-sdk/client-s3`, `sharp`, `@anthropic-ai/sdk`, `../src/lib/themes`, `../src/lib/utils`). Any "module not found" or syntax error will surface as an earlier different error.
-
-**Do NOT** use a dynamic-import or any test that loads the script with env vars present — `main()` runs at module load and will start writing to the DB. (Earlier draft of this plan had that bug; cost us 3 partial-row inserts that had to be hand-cleaned. Ask before "smoke testing" anything that imports this script.)
-
----
-
-### Task 4: Add npm script
-
-**Files:**
-- Modify: `package.json`
-
-- [ ] **Step 1: Add to scripts block**
-
-In `package.json`, locate the `"scripts"` block (after `"import:descriptions"`), add:
-
-```json
-"import:archive": "tsx --env-file=.env.local scripts/import-archive.ts"
-```
-
-The block (showing only the addition, preserve all existing scripts):
-
-```json
-"scripts": {
-  "dev": "next dev",
-  "build": "next build",
-  "start": "next start",
-  "lint": "next lint",
-  "import:csv": "tsx scripts/import-csv.ts",
-  "import:images": "tsx scripts/migrate-images.ts",
-  "seed:categories": "tsx scripts/seed-categories.ts",
-  "generate:descriptions": "tsx scripts/generate-descriptions.ts",
-  "import:descriptions": "tsx --env-file=.env.local scripts/import-human-descriptions.ts",
-  "import:archive": "tsx --env-file=.env.local scripts/import-archive.ts",
-  "migrate:all": "tsx --env-file=.env.local scripts/migrate-and-describe.ts",
-  "db:migrate": "tsx scripts/run-migration.ts"
-},
-```
-
-- [ ] **Step 2: Verify**
-
-Run: `npm run | grep import:archive`
-Expected: `import:archive`
-
----
-
-### Task 5: Sync `supabase/migrations/001_initial.sql`
-
-**Files:**
-- Modify: `supabase/migrations/001_initial.sql`
-
-- [ ] **Step 1: Add `kind` column to the categories CREATE TABLE**
-
-In `supabase/migrations/001_initial.sql`, find this block (around line 19-27):
-
-```sql
-CREATE TABLE categories (
-  id            UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-  name          TEXT NOT NULL,
-  slug          TEXT UNIQUE NOT NULL,
-  description   TEXT,
-  sort_order    INT DEFAULT 0,
-  ai_suggested  BOOLEAN DEFAULT false,
-  created_at    TIMESTAMPTZ DEFAULT now()
-);
-```
-
-Replace with:
-
-```sql
-CREATE TABLE categories (
-  id            UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-  name          TEXT NOT NULL,
-  slug          TEXT UNIQUE NOT NULL,
-  description   TEXT,
-  sort_order    INT DEFAULT 0,
-  ai_suggested  BOOLEAN DEFAULT false,
-  -- Discriminator for what kind of category this is. 'format' is the
-  -- existing AI-suggested taxonomy (Drawings, Paintings, etc.); 'theme'
-  -- is the controlled subject taxonomy added in 2026-04-23.
-  kind          TEXT CHECK (kind IN ('format', 'theme')),
-  created_at    TIMESTAMPTZ DEFAULT now()
-);
-```
-
-- [ ] **Step 2: Verify**
-
-Run: `grep -A 12 "CREATE TABLE categories" supabase/migrations/001_initial.sql`
-Expected: shows the updated CREATE TABLE block with the `kind` column.
-
----
-
-### Task 6: Commit Phase 1 code
-
-**Files:** none (git only)
-
-- [ ] **Step 1: Stage and commit**
-
-Run:
-```bash
-git add src/lib/themes.ts scripts/test-themes.ts scripts/import-archive.ts package.json supabase/migrations/001_initial.sql
-git status
-```
-
-Expected: 5 files staged. No untracked CSV artifacts (tmp/ is gitignored).
-
-- [ ] **Step 2: Create commit**
-
-Run:
-```bash
-git commit -m "$(cat <<'EOF'
-Add 1stDibs archive import script + theme taxonomy helper
-
-scripts/import-archive.ts reconciles the artworks DB with the 1stDibs
-picks CSV. For each row:
-  - SKU exists in DB:   upsert theme tags only (no metadata clobber)
-  - SKU not in DB:      insert artwork from CSV metadata, download
-                        image, upload R2 4-variant pipeline, generate
-                        AI alt text via Claude Vision, attach themes
-
-Themes are a controlled vocabulary of 8 (music, people, plants,
-animals, abstract, other, food, pop culture) stored in the existing
-categories table with the new kind='theme' discriminator. Theme
-normalization (clir/clear prefix stripping, case folding, dedup,
-fixed-set validation) is extracted to src/lib/themes.ts so Project B
-(theme + decade UI dropdowns) can reuse it. node:test coverage on the
-normalization function.
-
-Out of scope per spec: image validation for existing artworks,
-inactive marking for DB-only SKUs, metadata refresh on existing rows.
-
-Schema migration to add categories.kind was applied manually before
-this commit. supabase/migrations/001_initial.sql synced to match.
-
-Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
-EOF
-)"
-```
-
-Expected: commit succeeds.
-
----
-
-## Phase 2: Run + verify
-
-### Task 7: Pre-run sanity check
-
-**Files:** none (verification)
-
-- [ ] **Step 1: Re-confirm overlap counts**
-
-Run:
-```bash
-npx tsx --env-file=.env.local -e "
-import fs from 'fs';
-import { parse } from 'csv-parse/sync';
-import { createClient } from '@supabase/supabase-js';
-
-async function main() {
-  const c = createClient(process.env.NEXT_PUBLIC_SUPABASE_URL, process.env.SUPABASE_SERVICE_ROLE_KEY);
-  const rows = parse(fs.readFileSync('tmp/ADD TAGS TO CREATIVE GROWTH PUBLIC ARCHIVE - 1stdibs_clir_picks_2026-03-17.csv', 'utf-8'), {columns:true, skip_empty_lines:true, relax_quotes:true, relax_column_count:true});
-  const csvSkus = new Set(rows.map(r => (r.SKU || '').trim()).filter(Boolean));
-
-  let all = []; let off = 0;
-  while (true) {
-    const { data } = await c.from('artworks').select('sku').range(off, off+999);
-    if (!data || data.length === 0) break;
-    all.push(...data);
-    if (data.length < 1000) break;
-    off += 1000;
-  }
-  const dbSkus = new Set(all.map(r => r.sku).filter(Boolean));
-
-  const overlap = [...csvSkus].filter(s => dbSkus.has(s)).length;
-  const newOnly = [...csvSkus].filter(s => !dbSkus.has(s)).length;
-  console.log('CSV unique SKUs:', csvSkus.size);
-  console.log('DB unique SKUs: ', dbSkus.size);
-  console.log('Branch A (themes only):', overlap);
-  console.log('Branch B (insert + image + AI):', newOnly);
-}
-main();
-" 2>&1 | grep -v notice
-```
-
-Expected output ballpark (should match the earlier brainstorming numbers):
-```
-CSV unique SKUs: 2059
-DB unique SKUs:  2110
-Branch A (themes only): 909
-Branch B (insert + image + AI): 1150
-```
-
-If the Branch B count is dramatically larger or smaller than ~1,150: STOP and re-read the data with the user before running. The Branch B count drives all the Vision API spend and runtime.
-
----
-
-### Task 8: Run the import
-
-**Files:** none (execution; produces gitignored artifacts)
-
-- [ ] **Step 1: Kick off the run**
-
-Run: `npm run import:archive`
-
-Expected behavior:
-- Prints "Parsed N rows..." then "Fetched M DB artworks..."
-- Prints `[25/N], [50/N], ...` progress lines every 25 rows
-- Will take ~1-3 hours for the full 2,059 rows (most of the time spent on Branch B's image download + R2 upload + Vision call)
-- Resumable: if killed, can re-run; the checkpoint file `scripts/.archive-import-progress.json` lets it skip done rows
-
-If you see frequent failures (more than ~5% of attempts), stop and report to user — likely a config issue (R2, Anthropic, network).
-
-- [ ] **Step 2: Spot-check the log**
-
-After the run completes, run:
-```bash
-awk -F',' 'NR>1 {print $2}' tmp/archive-import-log_*.csv | sort | uniq -c
-```
-
-Expected: counts by branch. Roughly:
-- `existing_themed`: ~909
-- `inserted`: ~1,150 (or close — some Branch B rows may have failed pieces but still count as `inserted`)
-- `failed`: should be small (<50)
-
-Report the actual counts back to the user.
-
-- [ ] **Step 3: Spot-check a Branch A artwork (themes attached, nothing else changed)**
-
-Run:
-```bash
-npx tsx --env-file=.env.local -e "
-import { createClient } from '@supabase/supabase-js';
-const c = createClient(process.env.NEXT_PUBLIC_SUPABASE_URL, process.env.SUPABASE_SERVICE_ROLE_KEY);
-c.from('artworks').select('sku, title, alt_text, alt_text_long, description_origin, categories:artwork_categories(category:categories(name, kind))').eq('sku', 'RB 25').then(({data, error}) => {
-  if (error) { console.error(error); process.exit(1); }
-  data.forEach(r => console.log(JSON.stringify(r, null, 2)));
-});
-"
-```
-
-Expected: `RB 25` (or whatever Branch A SKU you pick) should have its existing `alt_text`, `alt_text_long`, `description_origin` UNCHANGED from before the run. The `categories` array should now include a row with `kind: 'theme'` if RB 25 was tagged in the CSV.
-
-- [ ] **Step 4: Spot-check a Branch B artwork (newly inserted)**
-
-Pick any SKU from the Branch B side of the CSV (e.g., one with `inserted` in the log):
-```bash
-npx tsx --env-file=.env.local -e "
-import { createClient } from '@supabase/supabase-js';
-const c = createClient(process.env.NEXT_PUBLIC_SUPABASE_URL, process.env.SUPABASE_SERVICE_ROLE_KEY);
-c.from('artworks').select('sku, title, image_url, alt_text, alt_text_long, description_origin, on_website, categories:artwork_categories(category:categories(name, kind))').eq('sku', 'DA 109').then(({data, error}) => {
-  if (error) { console.error(error); process.exit(1); }
-  data.forEach(r => console.log(JSON.stringify(r, null, 2)));
-});
-"
-```
-
-Expected: `DA 109` (or another known new SKU) shows:
-- `image_url` is an R2 URL
-- `alt_text` and `alt_text_long` are populated
-- `description_origin: 'ai'`
-- `on_website: true`
-- `categories` includes the relevant theme(s) from the CSV
-
-If `description_origin` is null and the log says `ai_status: failed`, that's an acceptable degraded outcome — the row is still inserted and the existing `generate-descriptions.ts` can backfill later.
-
----
-
-## Done
-
-At this point:
-- Existing 909 artworks (matched on SKU) have theme tags attached. Their other fields are untouched.
-- ~1,150 new artworks have been inserted with metadata, R2-hosted images, AI-generated alt text, and theme tags. They're live (`on_website = true`).
-- Categories table has 8 new `kind = 'theme'` rows.
-- Two timestamped artifacts in `tmp/`: the per-row log and the checkpoint file.
-
-Hand off to the user to:
-- Spot-check the public site (`/collection`) — should now have ~3,200+ visible artworks.
-- Decide what to do with the ~1,200 DB-only artworks (separate project: inactive marking).
-- Decide whether to re-run image validation on existing artworks (separate project).
-- Move on to Project B: theme + decade UI dropdowns on the collection page.
